InSync (major update)
No change to main logic of strategy. Adding a way to look for not as
extreme oversold conditions when the stock is above the EMA300. Also,
lowering the size of the move (extremeties) to help capture large
caps/less volatile stocks.  Adding RSI2 exit plot to help ID exits if
the price target doesnt materialize.